requirements.txt
Flask==2.0.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
pandas==1.3.3
python-dotenv==0.19.0
gunicorn==20.1.0
beautifulsoup4==4.12.2
schedule==1.2.1 